refactor: add index for components to make import clear for pages
