+++ components/atoms/SectionTitle/index.tsx
@@ -1,0 +1,1 @@
+export { default, SectionBigTitle, SubTitle } from "./SectionTitle";
