+++ components/molecules/Carousel/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./Carousel";
