+++ components/molecules/SectionWrapper/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./SectionWrapper";
