+++ pages/_app.tsx
@@ -1,7 +1,7 @@
 import "../styles/globals.scss";
 import type { AppProps } from "next/app";
 
-import Layout from "../components/organisms/Layout/Layout";
+import Layout from "../components/organisms/Layout";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
